Refactor day 05
The main method now produces results for days 1-5.

Sadly we can't build this for the Atari platform (using Cross
Mint) at present, for a few reasons:
- Cross Mint appears not to support uint64_t. It looks like the
compiler itself doesn't support  64-bit long long type, which I
guess isn't necessarily that surprising.
- Aome of the range/view functions like count_if and copy don't
appear to be supported by gcc 13 (and certainly not by the
Cross Mint version thereof). So we either need to rewrite stuff
not to use those functions, or just just get over the whole Atari
thing. Which would be a shame, but hey ho.