WIP
Reading commands. Not quite sure if this skip_whitespace()
function is the right thing to do. Really, we want a "skip
carriage return and newline chars" function. It feels a bit wrong
to just skeep whitespace characters and then assume that the next
token is definitely the one that we want, given that the input
is line oriented and the way we're doing it might mean that,
worst case, we lose aligned with line boundaries.

Maybe using getline instead would do the trick here and would
be closer to what we want to do.

What I _really_ want to do is read in the lines one by one
(using getline() I guess) and then use regexes to parse them.
Shame I haven't learned to do regexes in C++ yet...