Try non-recursive version that does a breadth-first search
Uses a FIFO queue to ensure that it really is breadth first.

Also prunes eny exit nodes (i.e. nodes where a beam exits the
grid) because I _believe_ that a beam entering the grid at such
a node can't hit more squares than a ray that _enters_ the
grid at that node. I'm not actually 100% convinced that this is
the case but it works for the input data that I have...

Note that this doesn't shave an awful lot of time off. Part 2
runs in 486 ms on my box, which is a modest improvement but still
not as fast as the solution in Day16.kt which just uses mutable
state with abandon.