Bigger optimisation
Every time we make a new guess, jump forward to a
point in time when a falling block blocks our current
path.

I didn't cound the number of iterations that this
needs. It's more than a binary search would need,
but not so much bigger that I'm motivated to re-write
it again.